+++ src/ctest/db/migrate.clj
@@ -13,21 +13,26 @@
     [ctest.config :as c]
     [ctest.db.init :as init]
     [ctest.db.crud :as crud]
-    [taoensso.nippy :as nippy]))
+    [taoensso.nippy :as nippy]
+    [clojure.string :as str]
+    [clojure.stacktrace :as st]
+    [ctest.common :as common]
+    [clojure.edn :as edn]))
 
 
 (defn table-set
   [db-conn]
   (set
     (jdbc/query db-conn
-      ["SELECT name FROM sqlite_master WHERE type='table'"] :row-fn (comp keyword :name))))
+      ["SELECT name FROM sqlite_master WHERE type='table'"]
+      {:row-fn (comp keyword :name)})))
 
 
 (defn table-columns
   [db-conn, table]
   (jdbc/query db-conn [(format "PRAGMA table_info(%s)" (name table))]
-    :row-fn (comp keyword :name),
-    :result-set-fn vec))
+    {:row-fn (comp keyword :name)
+     :result-set-fn vec}))
 
 
 (defn db-data
@@ -67,12 +72,10 @@
       (fn [_, table, rows]
         (when (contains? existing-tables table)
           (let [table-cols (table-columns db-conn, table)]
-            (reduce
-              (fn [_, row-data]
-                (jdbc/insert! db-conn, table, (select-keys row-data table-cols))
-                nil)
-              nil
-              rows))))
+            (jdbc/insert-multi! db-conn, table
+              (mapv #(select-keys % table-cols) rows))
+            ; return nothing
+            nil)))
       nil
       table-data-map)))
 
@@ -84,3 +87,86 @@
     (let [db-conn (db-connection db-filename)]
       (insert-table-data db-conn, table-data-map)
       nil)))
+
+
+
+(defn new-user-role
+  [old-role]
+  (when old-role
+    [old-role (keyword "role" (name old-role))]))
+
+
+(defn user-role-update
+  [user-roles]
+  (let [renaming-map (into {}
+                       (keep new-user-role)
+                       user-roles)]
+    (jdbc/with-db-transaction [t-conn (c/db-connection)]
+      (let [user-list (crud/read-users t-conn)]
+        (doseq [new-user (mapv #(update % :role (comp renaming-map edn/read-string)) user-list)]
+          (println "  UPDATING" (:username new-user))
+          (crud/update-user-role t-conn, new-user))))))
+
+
+(defn user-role-update-check
+  []
+  (when-let [user-roles (not-empty (crud/user-role-set))]
+    (when (some #(-> % namespace (not= "role")) user-roles)
+      ["user roles" (partial user-role-update user-roles)])))
+
+
+(defn status-encoding-update
+  [encodings]
+  (let [negative-import-encoding (c/import-negative-result)
+        {negative-encodings true, in-progress-encodings false} (group-by #(= negative-import-encoding %) encodings)
+        ne-count (count negative-encodings)
+        ipe-count (count in-progress-encodings)]
+    (if (and (<= ne-count 1) (<= ipe-count 1))
+      (let [renaming-map (cond-> {}
+                           (== ne-count 1) (assoc (first negative-encodings) c/db-encoding-negative)
+                           (== ipe-count 1) (assoc (first in-progress-encodings) c/db-encoding-in-progress))]
+        (println "  UPDATING" (->> renaming-map (map #(apply format "%s => %s" %)) (str/join ", ")))
+        (crud/update-status (c/db-connection), renaming-map))
+      (throw (Exception. (format "At most 1 encoding of each type expected! Found negativ = %s and in-progress = %s" negative-encodings, in-progress-encodings))))))
+
+
+(defn status-encoding-update-check
+  []
+  (when-let [encodings (-> (crud/status-encoding-set)
+                         (disj nil)
+                         not-empty)]
+    (when (not-every? c/valid-status-encoding? encodings)
+      ["status encodings" (partial status-encoding-update encodings)])))
+
+
+(defn qrcode-path-update
+  []
+  (crud/remove-qrcode-prefix "qrcodes/"))
+
+
+(defn qrcode-path-update-check
+  []
+  (when (crud/patients-with-qrcode-prefix? "qrcodes/")
+    ["qrcode path" qrcode-path-update]))
+
+(defn required-updates
+  []
+  (->> [user-role-update-check, status-encoding-update-check, qrcode-path-update-check]
+    (keep (fn [f] (f)))
+    vec
+    not-empty))
+
+
+(defn upgrade-if-needed
+  [config-file]
+  (when (c/read+setup-config config-file)
+    (when-let [required-updates-vec (required-updates)]
+      (doseq [[desc, update-fn] required-updates-vec]
+        (println "UPGRADE" desc "STARTED.")
+        (try
+          (update-fn)
+          (catch Throwable t
+            (println "UPGRADE" desc "FAILED:")
+            (st/print-cause-trace t)
+            (common/exit 1 (str "UPGRADE " desc " FAILED!"))))
+        (println "UPGRADE" desc "FINISHED.")))))