+++ src/ctest/runtime.clj
@@ -18,7 +18,6 @@
     [ctest.reporting :as report])
   (:import
     (org.eclipse.jetty.server Server AbstractConnector)
-    org.eclipse.jetty.server.ssl.SslSelectChannelConnector
     org.apache.log4j.PropertyConfigurator))
 
 
@@ -32,34 +31,6 @@
   keep-running)
 
 
-(defn use-only-tls
-  "For security reasons use only TLSv1.1 and TLSv1.2."
-  [^Server jetty]
-  (let [allowed-protocols (into-array String ["TLSv1.1" "TLSv1.2"])
-        forbidden-protocols (into-array String ["SSLv3" "TLSv1"])]
-    (doseq [^SslSelectChannelConnector con (->> jetty
-                                             .getConnectors
-                                             (filter #(instance? SslSelectChannelConnector %)))]
-      (doto (.getSslContextFactory con)
-        (.setIncludeProtocols allowed-protocols)
-        (.setExcludeProtocols forbidden-protocols)))))
-
-
-(defn use-forwarding
-  "Configure jetty to be used via forwarding."
-  [^Server jetty]
-  (report/info "Startup" "Configuring jetty to be used via forwarding.")
-  (doseq [^AbstractConnector con (.getConnectors jetty)]
-    (.setForwarded con true)))
-
-(defn configurator
-  [& fns]
-  (let [fns (remove nil? fns)]
-    (fn [jetty]
-      (doseq [f fns]
-        (f jetty)))))
-
-
 (defn- server-url
   [host, port, ssl? server-root]
   (let [proto (if ssl? "https" "http")]
@@ -67,15 +38,15 @@
       (format "%s://%s:%s" proto host port)
       (format "%s://%s:%s/%s" proto host port server-root))))
 
+
 (defn start-server
   [app]
   (when-not running-server
     (alter-var-root #'keep-running (constantly true))
-    (let [{:keys [ssl?, forwarded?] :as config} (-> (c/server-config) (assoc :join? false)),
+    (let [{:keys [ssl?] :as config} (-> (c/server-config) (assoc :join? false)),
           server (jetty/run-jetty
                    app
                    (cond-> config
-                     (or ssl? forwarded?) (assoc :configurator (configurator (when ssl? use-only-tls), (when forwarded? use-forwarding)))
                      (not ssl?) (dissoc :ssl-port :key-password :keystore)))]
       (alter-var-root #'running-server (constantly server))
       (let [{:keys [host, port, ssl-port, ssl?, server-root]} (c/server-config)]
@@ -137,24 +108,6 @@
     (catch Throwable t
       (log/errorf "Exception when initiating server shutdown:\n%s"
         (report/cause-trace t)))))
-
-
-
-(defn configure-logging
-  "Configures the logging for ctest. Log level and log file can be specified in the configuration."
-  [{:keys [log-level, log-file] :as config}]
-  (let [props (doto (System/getProperties)
-                (.setProperty "log4j.rootLogger" (format "%s, file" (-> log-level name str/upper-case)))
-                (.setProperty "log4j.appender.file" "org.apache.log4j.RollingFileAppender")
-                (.setProperty "log4j.appender.file.File" (str log-file))
-                (.setProperty "log4j.appender.file.MaxFileSize" "4MB")
-                (.setProperty "log4j.appender.file.MaxBackupIndex" "5")
-                (.setProperty "log4j.appender.file.layout" "org.apache.log4j.PatternLayout")
-                (.setProperty "log4j.appender.file.layout.ConversionPattern" "%d{yyyy.MM.dd HH:mm:ss} %5p %c: %m%n")
-                ; jetty is too chatty
-                (.setProperty "log4j.logger.org.eclipse.jetty" "INFO"))]
-    (PropertyConfigurator/configure props))
-  nil)
 
 
 (defn- signal-available?