+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# Version 0.3.0
+
+* Upgrade to Clojure 1.10.1.
+
+* Upgrade of all dependencies.
+
+* Tested with Java 8, Java 11 and Java 14.
+
+* Removes configuration parameter `:forwarded?`.
+
+* Switched `import` task to batched insert.
+
+* Switched CSV import to batched update.
+
+* Default configuration sets `:ssl? false`.
+
+* Fixes shutdown logic in import loop.
+
+* New `upgrade` task to migrate from `0.1.3` and `0.2.x` to `0.3.x`. 
+
 # Version 0.2.2
 
 * Fixes a bug in user management dialog.